moving pose overlay related stuff from FrameWidget to FrameWidgetWithInteractiveOverlays
